requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
Pillow==7.0.0
pycparser==2.19
pylint==2.4.4
PySocks==1.7.1
python-telegram-bot==12.3.0
PyYAML==5.3
scipy==1.4.1
six==1.14.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tornado==6.0.3
wrapt==1.11.2
